Address SWE-bench review findings
