+++ src/inference_endpoint/evaluation/swebench_service/swebench_service/MINI_SWE_AGENT_LICENSE.md
